+++ src/main/java/hive/WorkerBee.java
@@ -1,7 +1,5 @@
 package hive;
 
-import com.sun.corba.se.spi.orbutil.threadpool.Work;
-
 import java.util.Random;
 import java.util.logging.Logger;