+++ CHANGELOG.md
@@ -4,7 +4,7 @@ NB: All notable changes to this project will be documented in this file.  The fo
 
 ## Week of 2019-10-28
 - Repo tests now force serialized access
-- Wrote a new service layer tests that demonstrates parallel calls even with single-threaded executor
+- Wrote a new future.service layer tests that demonstrates parallel calls even with single-threaded executor
 - Restructured tests so that we can share code between them.
 - Added a change log :)
 - Commented for clarity throughout project